refactor(src): Convert to TypeScript
