+++ src/index.ts
@@ -1,8 +1,7 @@
-// LICENSE : MIT
 "use strict";
 import RuleHelper from "./textlint-rule-helper";
 import IgnoreNodeManager from "./IgnoreNodeManager"
-module.exports = {
+export {
     IgnoreNodeManager,
     RuleHelper
 };
